osd progress bar tweaks (+color follow)
made the OSD slider button a little larger and slightly translucent so
you can see the progress bar under it as you move it - also it looks a
little slightly off center but with a mouse or finger you will need to
grab it offset from dead center to see where you are at while moving.

feel free to revert, suggest or make changes.

ALSO - should the color of the slider button and progress bar follow
buttonfocus or mediafocus skin setting colors?